feat/#56: Entire viewmodel 분리
